Resolved merge conflict issues resulting in compile time failure.

+++ client-api/src/main/java/com/talentica/hungryHippos/client/job/Job.java
@@ -10,5 +10,5 @@ public interface Job {
     int[] getDimensions();
     int getPrimaryDimension();
     int getIndex();
-    long getMemoryFootprint(int rowCount);
+	long getMemoryFootprint(long rowCount);
 }
